[BUG#290] 검색 기록 오류 해결
